- Reduce duplication in v4 v5 cltu definitions
